[feat] 프래그먼트 다시 합치기
