+++ app/src/main/java/org/sopt/navermap/presentation/main/MainsearchFragment.kt
@@ -2,7 +2,6 @@ package org.sopt.navermap.presentation.main
 
 import android.os.Bundle
 import android.view.View
-import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.viewModels
 import org.sopt.dosopttemplate.util.binding.DataBindingFragment
 import org.sopt.navermap.R
@@ -15,18 +14,14 @@ class MainsearchFragment :
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        val locationGridAdapter = LocationGridAdapter(requireContext())
+        binding.rvMainSearchGrid.adapter = locationGridAdapter
+        locationGridAdapter.setLocationList(viewModel.mockLocationList.subList(0, 4))
 
-        val currentFragment = childFragmentManager.findFragmentById(R.id.fcv_main_search_result_grid)
-        if (currentFragment == null ) {
-            childFragmentManager.beginTransaction()
-                .add(R.id.fcv_main_search_result_grid, MainsearchgridFragment())
-                .commit()
-        }
-        val secondCurrentFragment = childFragmentManager.findFragmentById(R.id.fcv_main_search_result_list)
-        if (currentFragment == null ) {
-            childFragmentManager.beginTransaction()
-                .add(R.id.fcv_main_search_result_list, MainsearchlistFragment())
-                .commit()
-        }
+        val locationListAdapter = LocationListAdapter(requireContext())
+        binding.rvMainSearchList.adapter = locationListAdapter
+        locationListAdapter.setLocationList(viewModel.mockLocationList)
+
+
     }
 }
