--- app/src/main/res/layout/fragment_main_search_list.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    >
-    <data
-        >
-        <import type="android.view.View"/>
-        <variable
-            name="viewModel"
-            type="org.sopt.navermap.presentation.main.MainActivityViewModel" />
-    </data>
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_main_search_list"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/v_search_mid_bar" />
-
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
